feat: Enhance auction joining functionality with additional parameters
- Added auctionStartRate and currentBidders to JoinTheAuction widget.
- Updated CounterView to utilize auctionStartRate and currentBidders for calculations.
- Improved error handling in WebViewJoinToAuctionScreen for server responses.
- Refactored CompleteShippingInformationScreen to use a StatefulWidget for better state management and user data fetching.
- Updated invoice details text for clarity.
- Enhanced account details screen to fetch and display user phone code dynamically.
- Implemented singleton pattern for WebSocketCubit to manage WebSocket connections more efficiently.
- Added server time handling and improved error management in WebSocket communication.
- Integrated WebSocketCubit into the main application for real-time updates.

+++ lib/feature/home/home_details/notstart/data/model/qadim_auction_response.dart
@@ -62,7 +62,7 @@ class NotstartAuction {
   final int? auctionDurationMinutes;
 
   @JsonKey(name: 'auction_start_rate')
-  final int? auctionStartRate;
+  final double? auctionStartRate;
 
   @JsonKey(name: 'product_sku')
   final String? productSku;
