+++ lib/feature/home/home_details/ready/data/model/sayantaliq_auction_response.dart
@@ -58,7 +58,7 @@ class ReadyAuction {
   @JsonKey(name: 'auction_duration_minutes')
   final int? auctionDurationMinutes;
   @JsonKey(name: 'auction_start_rate')
-  final int? auctionStartRate;
+  final double? auctionStartRate;
   @JsonKey(name: 'product_sku')
   final String? productSku;
   final bool? isRegister;
